chore: lint and spacing changes fix: fix misspellings
