+++ src/services/neo4j.js
@@ -15,10 +15,7 @@ module.exports = async () => {
         if (!await driver.verifyAuthentication()) throw ({ message: "Auth neo4j error" });
         const session = driver.session();
 
-        const neo4j = {
-            driver,
-            session
-        };
+        const neo4j = { driver, session };
 
         return neo4j;
     } catch (err) {
